liaison front + add service

+++ src/main/java/com/example/projet/projet/Repository/CategorieRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.projet.projet.Model.CategorieEntity;
 
-public interface CategorieRepository extends CrudRepository<CategorieEntity,Long> {
+public interface CategorieRepository extends CrudRepository<CategorieEntity,Integer> {
     
 }
