+++ src/main/java/com/example/projet/projet/Repository/RoleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.projet.projet.Model.RoleEntity;
 
-public interface RoleRepository extends CrudRepository<RoleEntity,Long> {
+public interface RoleRepository extends CrudRepository<RoleEntity,Integer> {
     
 }
