+++ src/main/java/com/example/projet/projet/Repository/TechnicienRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.example.projet.projet.Model.TechnicienEntity;
 
-public interface TechnicienRepository extends CrudRepository<TechnicienEntity,Long> {
+public interface TechnicienRepository extends CrudRepository<TechnicienEntity,Integer> {
     
 }
